% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrologit_methods.R
\name{print.rcrologit}
\alias{print.rcrologit}
\title{Print Method for Random Coefficients Rank-Ordered Logit}
\usage{
\method{print}{rcrologit}(x, printFE = FALSE, ...)
}
\arguments{
\item{x}{Class "rcrologit" object, obtained by calling \code{\link{rcrologit}}.}

\item{printFE}{Whether fixed effects coefficients and standard errors should be displayed.}

\item{...}{Other arguments.}
}
\value{
No return value, called to print \code{\link{rcrologit}} results.
}
\description{
The print method for for random coefficients rank-ordered logit fitted objects.
}
\seealso{
\code{\link{rcrologit}} for estimation and inference for random coefficients rank-ordered logit.

Supported methods: \code{\link{print.rcrologit}}.
}
\author{
Chiara Motta, University of California Berkeley. \email{cmotta@berkeley.edu}

Filippo Palomba, Princeton University. \email{fpalomba@princeton.edu}
}
