% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrologit-package.R
\docType{package}
\name{rcrologit-package}
\alias{rcrologit-package}
\alias{_PACKAGE}
\title{\code{rcrologit}: A Package to Estimate Random Coefficients Rank-Ordered Logit Models.}
\description{
The package implements estimation and inference procedures for random coefficient rank-ordered logit models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/filippopalomba/rcrologit}
  \item Report bugs at \url{https://github.com/filippopalomba/rcrologit/issues}
}

}
\author{
Chiara Motta, University of California Berkeley. \email{cmotta@berkeley.edu}

Filippo Palomba, Princeton University. \email{fpalomba@princeton.edu}
}
