% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrologit.R
\name{rcrologit}
\alias{rcrologit}
\title{Estimation and Inference for Random Coefficients Rank-Ordered Logit}
\usage{
rcrologit(
  dataprep,
  Sigma = "diagonal",
  S = 50L,
  approx.method = "MC",
  stdErr = "numerical",
  Ncores = 1L,
  seed = 8894L,
  verbose = FALSE,
  control.opts = NULL
)
}
\arguments{
\item{dataprep}{object of class 'rcrologit' prepared via \code{\link{dataPrep}}.}

\item{Sigma}{structure of the variance-covariance of the random coefficients. It must be one of
"diagonal" or "cholesky". Default is \code{Sigma="diagonal"}. See \strong{Details} section for more.}

\item{S}{integer denoting the number of simulations when approximating the integrals in the conditional
choice probabilities. Default is \code{S=50}.}

\item{approx.method}{string indicating the procedure to approximate the integrals in the conditional
choice probabilities when including random coefficients.
At the moment only approximation via monte-carlo simulation is available. Future releases will include
importance sampling and other alternatives}

\item{stdErr}{string denoting whether standard error should be estimated and how. Available options are:
\itemize{
\item{\emph{analytical}: (default) which uses
the closed form of the variance of the score and the hessian of the likelihood to estimate standard
errors robust to misspecification}
\item{\emph{analytical - norobust}: which uses
the closed form of the variance of the score to estimate standard errors. Note that this standard errors are not
robust to misspecification}
\item{\emph{numerical}: which numerically approximates the Hessian matrix. Note that this standard errors are not
robust to misspecification. Default for random coefficients rologit at the moment.}
\item{\emph{skip}: if no standard errors have to be computed.}
}}

\item{Ncores}{an integer indicating the number of cores to be used in simulating conditional choice probabilities.
It affects speed only when random coefficients are included in the model. Speed gains are sensible whenever \eqn{S\geq 100}.}

\item{seed}{an integer setting the seed when \code{Ncores} is larger than 1. This is passed to the underlying
\code{parallel::mclapply} to ensure replicability of results.}

\item{verbose}{if \code{TRUE} prints additional information in the console.}

\item{control.opts}{a list containing options to be passed to the underlying optimizer \code{optim}.}
}
\value{
The function returns a list containing the following objects:
\item{b}{vector containing all the estimated parameters}
\item{bfix}{vector containing the estimated parameters corresponding to the "fixed" coefficients}
\item{bhet}{vector containing the estimated parameters corresponding to the "random" coefficients}
\item{Lambda}{loading matrix of the shocks to the random coefficients}
\item{Sigma}{variance-covariance matrix of the estimated parameters}
\item{param.spec}{a list containing some parameters describing the specification chosen by the user}
}
\description{
This function takes the object prepared by \code{dataPrep} and estimates a
random coefficient rank-ordered logit. The rank-ordered logit - sometimes termed \emph{exploded logit model} - was
originally proposed in \insertCite{beggs1981AssessingPotentialDemand;textual}{rcrologit} and it is an extension of the
\insertCite{luce1959IndividualChoiceBehavior;textual}{rcrologit}-\insertCite{mcfadden1974FrontiersEconometrics;textual}{rcrologit} model.

These models automatically implies independence of irrelevant alternatives \insertCite{debreu1960ReviewRDLuce}{rcrologit}. However,
\insertCite{mcfadden2000MixedMNLModels;textual}{rcrologit} show that if agents are allowed to have heterogeneous
tastes (i.e., random coefficients), then the conditional probability choices induced by the latent utility model can
match those implied by virtually any discrete choice probability model.

The package \code{\link{rcrologit}}, depending on the type of covariates specified in \code{dataPrep},
allows the user to estimate:
\itemize{
\item{the standard rank-ordered logit model if either \code{covs.fix} or \code{covsInt.fix} are specified.}
\item{the random coefficients rank-ordered logit model if either \code{covs.het} or \code{covsInt.het} are specified.}
\item{the mixed random coefficients rank-ordered logit model if at least one of \code{covs.fix} or \code{covsInt.fix} and
at least one of \code{covs.het} or \code{covsInt.het} are specified.}
}

For more information on the underlying specification see the \href{https://github.com/filippopalomba/rcrologit}{official repository}.
}
\details{
\itemize{
\item{\strong{Variance-Covariance of random parameters.} The option \code{Sigma} allows the user to
model directly the covariance structure of the random coefficients. Precisely, it shape \eqn{\boldsymbol{\Sigma}} in
\deqn{\left[\begin{array}{c}\boldsymbol{\alpha}_i \\ \boldsymbol{\beta}_i \end{array}\right]\sim
\mathsf{N}\left(\left[\begin{array}{c}\boldsymbol{\alpha}_{\mathtt{R}} \\ \boldsymbol{\beta}_{\mathtt{R}}\end{array}\right],
\boldsymbol{\Sigma}\right).}
In practice, to avoid issues with the positive-definiteness of \eqn{\boldsymbol{\Sigma}} during the optimization routine,
it is more convenient to model the random coefficients as
\deqn{\left[\begin{array}{c}\boldsymbol{\alpha}_i \\ \boldsymbol{\beta}_i \end{array}\right]=
\left[\begin{array}{c}\boldsymbol{\alpha}_{\mathtt{R}} \\ \boldsymbol{\beta}_{\mathtt{R}}\end{array}\right] + \Lambda \mathbf{u}_i,
\quad \mathbf{u}_i\sim\mathsf{N}(0,\boldsymbol{I}).}
The option \code{Sigma} directly models the shape of \eqn{\Lambda} and allows the user to choose between a diagonal structure
(\code{Sigma="diagonal"}) and a lower-triangular structure (\code{Sigma="cholesky"}).
}
\item{\strong{Model.}
The function estimates a random coefficient rank-ordered logit model induced by the latent utility model
\insertCite{mcfadden1974FrontiersEconometrics;textual}{rcrologit}
\deqn{U_{i\ell} = u_{i\ell} + \epsilon_{i\ell},\quad i=1,2,\ldots,n,\quad j=0,1,\ldots,J.}
In its most general form, we model \eqn{u_{i\ell}} as
\deqn{u_{i\ell}=X_{i\ell}^\top\boldsymbol{\beta}_{\mathtt{F}} + Z_i^\top\boldsymbol{\alpha}_{\ell,\mathtt{F}} +
W_{i\ell}^\top\boldsymbol{\beta}_i + V_i^\top\boldsymbol{\alpha}_{i\ell} + \delta_\ell}
where
\itemize{
\item{\eqn{X_{i\ell}} are covariates varying at the unit-alternative level whose coefficients are modelled as fixed.
The user can specify these covariates via the option \code{covs.fix}.}
\item{\eqn{Z_{i}} are covariates varying at the unit level whose coefficients are modelled as fixed.
The user can specify these covariates via the option \code{covsInt.fix}. These coefficients are interacted with
a dummy for the choice and treated as alternative varying covariates,
i.e., \eqn{Z_{i\ell}=\sum_{j=1}^JZ_i\times\mathbf{1}(j=\ell)}, where $J=0$ is the reference group.}
\item{\eqn{W_{i\ell}} are covariates varying at the unit-alternative level whose coefficients are modelled as random.
The user can specify these covariates via the option \code{covs.het}.}
\item{\eqn{V_{i}} are covariates varying at the unit level whose coefficients are modelled as random.
The user can specify these covariates via the option \code{covsInt.het}. These coefficients are interacted with
a dummy for the choice and treated as alternative varying covariates,
i.e., \eqn{V_{i\ell}=\sum_{j=1}^JV_i\times\mathbf{1}(j=\ell)}, where $J=0$ is the reference group.}
\item{the random coefficients}{ are modeled as a joint multivariate normal and are i.i.d. across units,
\deqn{\left[\begin{array}{c}\boldsymbol{\alpha}_i \\ \boldsymbol{\beta}_i \end{array}\right]\sim
\mathsf{N}\left(\left[\begin{array}{l}\boldsymbol{\alpha}_{\mathtt{R}} \\
\boldsymbol{\beta}_{\mathtt{R}}\end{array}\right],\boldsymbol{\Sigma}\right) }}
\item{\eqn{\delta_\ell}}{ are alternative-specific fixed effects that can be specified via the option \code{FE}.}
\item{\eqn{\epsilon_{i\ell}\sim\mathsf{Gu}(0,1)}}{are idiosyncratic i.i.d. shocks.}
}
The parameter vector to be estimated is thus
\deqn{\theta = \left(\boldsymbol{\beta_\mathtt{F}}^\top,\boldsymbol{\beta_\mathtt{R}}^\top,
\boldsymbol{\alpha_\mathtt{F}}^\top,\boldsymbol{\alpha_\mathtt{R}}^\top,
\mathrm{vech}(\boldsymbol{\Sigma})^\top,\{\delta\}_{j=1}^J\right)^\top,}
where the first alternative-fixed effect has been normalized to 0.}
}

For more information on the underlying specification see the \href{https://github.com/filippopalomba/rcrologit}{official repository}.
}
\examples{
data <- rcrologit_data
dataprep <- dataPrep(data, idVar = "Worker_ID", rankVar = "rank",
                     altVar = "alternative",
                     covsInt.fix = list("Gender"),
                     covs.fix = list("log_Wage"), FE = c("Firm_ID"))

rologitEst <- rcrologit(dataprep)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{dataPrep}}
}
\author{
Chiara Motta, University of California Berkeley. \email{cmotta@berkeley.edu}

Filippo Palomba, Princeton University. \email{fpalomba@princeton.edu}
}
